+++ src/DIC/NotifyUsersDialog.cs
@@ -2,14 +2,14 @@
 using System.Collections.Generic;
 using System.Threading;
 using System.Threading.Tasks;
-using Bot.Services.Reminds;
+using Bot.States;
 using Microsoft.Bot.Builder;
 using Microsoft.Bot.Builder.Dialogs;
 using Microsoft.Bot.Builder.Integration.AspNet.Core;
 using Microsoft.Bot.Schema;
 using Microsoft.Extensions.Configuration;
 
-namespace Bot.Clockify
+namespace Bot.DIC
 {
     public class NotifyUsersDialog : ComponentDialog
     {
