+++ src/Services/Reminds/GenericRemindService.cs
@@ -1,7 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Threading;
 using System.Threading.Tasks;
 using Bot.States;
 using Microsoft.Bot.Builder;
