+++ src/Services/Reminds/SmartWorkingRemindService.cs
@@ -2,6 +2,7 @@
 using System.Threading.Tasks;
 using Bot.Clockify;
 using Bot.Dialogs;
+using Bot.States;
 using Microsoft.Bot.Builder;
 using Microsoft.Bot.Builder.Dialogs;
 using Microsoft.Extensions.Configuration;
