+++ src/Startup.cs
@@ -10,6 +10,7 @@ using Bot.Security;
 using Bot.Services;
 using Bot.Services.Reminds;
 using Bot.Services.Reports;
+using Bot.States;
 using Bot.Supports;
 using Bot.Utils;
 using F23.StringSimilarity;
