+++ src/States/UserProfilesProvider.cs
@@ -1,17 +1,11 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Linq;
-using System.Threading;
 using System.Threading.Tasks;
 using Bot.Data;
-using Bot.States;
 using Microsoft.Bot.Builder;
 using Microsoft.Bot.Builder.Azure;
-using Microsoft.Extensions.Configuration;
-using Microsoft.WindowsAzure.Storage;
-using Microsoft.WindowsAzure.Storage.Blob;
 
-namespace Bot.Services.Reminds
+namespace Bot.States
 {
     public class UserProfilesProvider : IUserProfilesProvider
     {
